refactor(functions): align delivery and fixture concepts
