--- packages/conformance/src/functions-rtdb-fixture.ts
@@ -1,16 +1,0 @@
-export { functionsRtdbRows } from '../registry/functions-rtdb.ts';
-export {
-  probe as exactProbe,
-} from '../probes/functions-rtdb/functions-rtdb-onvaluecreated-exact-create.ts';
-export {
-  probe as startupProbe,
-} from '../probes/functions-rtdb/functions-rtdb-onvaluecreated-startup-existing.ts';
-export {
-  probe as wildcardProbe,
-} from '../probes/functions-rtdb/functions-rtdb-onvaluecreated-wildcard-batches.ts';
-export {
-  probe as descendantProbe,
-} from '../probes/functions-rtdb/functions-rtdb-onvaluecreated-descendant-projection.ts';
-export {
-  probe as failureProbe,
-} from '../probes/functions-rtdb/functions-rtdb-onvaluecreated-failed-execution.ts';
